bugfix: unsigned dummy values were always 0

+++ dummy_creator.cpp
@@ -19,13 +19,15 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 */
 
-// $Revision: 13252 $ $Date:: 2020-06-12 #$ $Author: serge $
+// $Revision: 13274 $ $Date:: 2020-06-16 #$ $Author: serge $
 
 #include "dummy_creator.h"
 
 #include <limits>          // std::numeric_limits<T>::max
 #include <cstdlib>         // rand()
 #include <algorithm>       // std::generate_n
+
+//#include <iostream>      // DEBUG
 
 namespace basic_parser
 {
@@ -36,7 +38,9 @@ namespace dummy
 template<class T>
 T get_rand()
 {
-    auto res = static_cast<T>( std::rand() % static_cast<int>( std::numeric_limits<T>::max() ) );
+    // DEBUG: std::cout << "limit " << std::numeric_limits<T>::max() << " <casted>"<< static_cast<std::uint64_t>( std::numeric_limits<T>::max() ) << "\n";
+
+    auto res = static_cast<T>( std::rand() % static_cast<std::uint64_t>( std::numeric_limits<T>::max() ) );
 
     return res;
 }
